% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_.R
\name{aggregate_}
\alias{aggregate_}
\title{Estimate the second stage of the Fama-MacBeth regression.}
\usage{
aggregate_(model, l_fit)
}
\arguments{
\item{model}{A list specifying the model.
E.g., x and y fields should contain rhs and lhs of the model.}

\item{l_fit}{A list of "fits" from the first stage.
Each "fit" is a list with the result for that "date".}
}
\value{
A list with the results of the Fama-MacBeth regression.
        This list is a fmreg object.
}
\description{
\code{aggregate_} aggregates the results from the first stage of
the Fama-MacBeth regression.
}
\seealso{
\code{\link[fmreg]{estimate_}}
}
\keyword{internal}

