% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_na_drop.R
\name{drop_na}
\alias{drop_na}
\title{Drop 'NA's.}
\usage{
drop_na(.data, filter_vars)
}
\arguments{
\item{.data}{Data frame with the data.}

\item{filter_vars}{Names of the variables
for which NA cases should be deleted.}
}
\description{
\code{drop_na} deletes NA cases in the specified vars from a data frame.
}
\keyword{internal}

