% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_.R
\name{load_data_}
\alias{load_data_}
\title{Load data for specified date.}
\usage{
load_data_(.data, .date, date_var)
}
\arguments{
\item{.data}{"Data frame - like" structure.
\code{.data} can be a data frame, a data table,
or a tibble, including a "connection" to
a out-of-memory database.}

\item{.date}{Date that is used to filter the data.}

\item{date_var}{Name of the date variable (should be a colname in .data)}
}
\value{
Part (rows) of \code{.data} that corresponds to \code{.date}.
}
\description{
\code{load_data_} takes only part of the data that corresponds to
a specified date.
}
\keyword{internal}

