% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fmb_vars.R
\name{make_fmb_vars}
\alias{make_fmb_vars}
\title{Make Fama-MacBeth variables.}
\usage{
make_fmb_vars(df_data, vars, date_var, id_var, cutoffs = c(0.01, 0.99),
  method = c("winsorize", "trim")[1])
}
\arguments{
\item{df_data}{Data frame with the data.}

\item{vars}{Character vector with names of variables
that are used to construct fmb_vars.}

\item{date_var}{A name of the variable with the dates (time dimension).}

\item{id_var}{A name of the variable with the IDs (e.g., `permno`).}

\item{cutoffs}{A vector with two cutoffs (lower and upper)
for winsorization or trimming.
E.g., c(0.01, 0.99) means 1\% and 99\% levels cutoffs.}

\item{method}{A name of the method to use: "winsorize" or "trim".

Note: The function rearranges the data
      by sorting on id_var and then on date_var.
      The function deletes all rows with NAs in fmb_vars.}
}
\description{
Makes Fama-MacBeth variables for given names of variables
and adds them to the data frame.
}

