% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_tails_.R
\name{modify_tails_}
\alias{modify_tails_}
\title{Modify tails of a vector.}
\usage{
modify_tails_(x, cutoffs, method = c("winsorize", "trim")[1])
}
\arguments{
\item{x}{A vector that should be modified.}

\item{cutoffs}{A vector with cutoffs.
E.g., cutoffs = c(0.01, 0.99) means
      "winsorize/trim at 1\% and 99\% levels".}

\item{method}{Name of the modification method to be applied to the vector.
Supports two methods: "winsorize" and "trim".}
}
\description{
\code{modify_tails_} modify a vector by winsorizing or trimming the tails.
}
\keyword{internal}

