% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_prepare_data_.R
\name{prepare_data_}
\alias{prepare_data_}
\title{Prepare data for the Fama-MacBeth regression.}
\usage{
prepare_data_(.data, y, X, date_var, winsorize, trim, cutoffs)
}
\arguments{
\item{.data}{Data frame with the data}

\item{y}{Name of the dependent variable}

\item{X}{Names of the regressors}

\item{date_var}{Name of the date variable}

\item{winsorize}{TRUE/FALSE}

\item{trim}{TRUE/FALSE}

\item{cutoffs}{Lower and upper cutoffs for winsorization/trimming}
}
\value{
data frame with the transformed data.
}
\description{
\code{prepare_data_} transforms the data before running
the Fama-MacBeth regression.
}
\keyword{internal}

