% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_from_lm_fit_.R
\name{r_squared_from_lm_fit_}
\alias{r_squared_from_lm_fit_}
\title{R^2 from \code{lm.fit} output.}
\usage{
r_squared_from_lm_fit_(obj, intercept = TRUE, is_adj = FALSE)
}
\arguments{
\item{obj}{List: output from \code{lm.fit}.}

\item{intercept}{Logical: does the model contain an intercept?}

\item{is_adj}{Logical: do you need the adjusted R^2?}
}
\description{
\code{r_squared_from_lm_fit_} calculates R^2 (plain or adjusted)
from the output of \code{lm.fit}.
}
\keyword{internal}

