% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_trim.R
\name{trim}
\alias{trim}
\title{Trim.}
\usage{
trim(x, cutoffs)
}
\arguments{
\item{x}{A vector that should be trimmed (truncated).}

\item{cutoffs}{A vector with cutoffs.
E.g., cutoffs = c(0.01, 0.99) means
      "trim at 1\% and 99\% levels".}
}
\value{
A trimmed vector of the same length as x.

The values below and above the cutoffs are filled with NAs.
}
\description{
Trims (truncates) a vector at given levels of cutoffs.
}

