% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize.}
\usage{
winsorize(x, cutoffs)
}
\arguments{
\item{x}{A vector that should be winsorized.}

\item{cutoffs}{A vector with cutoffs.
E.g., cutoffs = c(0.01, 0.99) means
      "winsorize at 1\% and 99\% levels".}
}
\description{
Winsorizes a vector at given levels of cutoffs.
}

