% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmbreg.R
\name{fmbreg}
\alias{fmbreg}
\title{Fama-MacBeth regression.}
\usage{
fmbreg(.data, y, X, date_var, intercept = TRUE)
}
\arguments{
\item{.data}{Data frame with the data}

\item{y}{Name of the dependent variable
(should be a colname in .data)}

\item{X}{Names of the regressors
(should be colnames in .data)}

\item{date_var}{Name of the date variable
(should be a colname in .data)}

\item{intercept}{Logical/Double: If FALSE, no intercept.
If TRUE, add intercept.
If intercept = a,
the constant regressor is of form
rep(a, times = N) instead of rep(1L, times = N).}
}
\value{
A list: $est      - vector of estimates of the coefficients;
                $t_stat   - vector of the t-stats for the coefficients;
                $cs_estimates - data frame with cross-sectional estimates
                                for every period.
}
\description{
\code{fmbreg} estimates Fama-MacBeth regression.
}

