% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmreg_.R
\name{fmreg_}
\alias{fmreg_}
\title{Fama-MacBeth regression.}
\usage{
fmreg_(.data, y, X, date_var, intercept = TRUE, winsorize = FALSE,
  trim = FALSE, cutoffs = c(0.01, 0.99), min_obs = 100)
}
\arguments{
\item{.data}{Data frame with the data}

\item{y}{Name of the dependent variable
(should be a colname in .data)}

\item{X}{Names of the regressors
(should be colnames in .data)}

\item{date_var}{Name of the date variable
(should be a colname in .data)}

\item{intercept}{Logical/Double: If FALSE, no intercept.
If TRUE, add intercept.
If intercept = a,
the constant regressor is of form
rep(a, times = N) instead of 
rep(1L, times = N).}

\item{winsorize}{Logical: If TRUE, winsorize the regressors.}

\item{trim}{Logical: If TRUE, trim the regressors.}

\item{cutoffs}{Vector with two numbers between 0 and 1:
lower and upper cutoffs for winsorization/trimming.}

\item{min_obs}{Number: If a cross-section contains less than
\code{min_obs} observations, a warning is issued.}
}
\value{
A list: $fm_estimates - data frame with Fama-MacBeth estimates;
                $cs_estimates  - data frame with cross-sectional estimates
                                 for every period.
}
\description{
\code{fmreg_} estimates Fama-MacBeth regression.
}
\seealso{
\code{\link[fmreg]{fmreg}}
}

